Something is messed up with git on this pc?
