+++ Application/src/main/java/com/apMadden/flowTestApp/view/BaseViewModel.kt
@@ -1,5 +1,6 @@
 package com.apMadden.flowTestApp.view
 
+import android.arch.lifecycle.ViewModel
 import android.view.KeyEvent
 import android.view.inputmethod.EditorInfo
 import android.widget.TextView
@@ -9,7 +10,7 @@ import android.widget.TextView
 /**
  * Created by madde on 12/23/2017.
  */
-open class BaseViewModel {
+open class BaseViewModel : ViewModel() {
     lateinit var focusChangeListener: FocusChangeListener
     private var actionListener: EditTextActionListener? = null
