+++ Application/src/main/res/layout/activity_camera.xml
@@ -18,5 +18,4 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#000"
-    tools:context="com.apMadden.android.conversationalPrototype.CameraActivity" />
+    android:background="#000" />
