+++ Application/src/main/res/layout/activity_main.xml
@@ -13,9 +13,24 @@
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
+        <!--
+        <android.support.design.widget.AppBarLayout
+            android:id="@+id/appbar_checkout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+
+            <android.support.v7.widget.Toolbar
+                android:id="@+id/toolbar_checkout"
+                layout="@layout/widget_toolbar"
+                android:layout_width="match_parent"
+                android:layout_height="?attr/actionBarSize"
+                app:layout_scrollFlags="scroll|enterAlways"/>
+        </android.support.design.widget.AppBarLayout>-->
+
         <android.support.v7.widget.RecyclerView
             android:id="@+id/rv_main_list"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
+            android:layout_height="wrap_content" />
     </android.support.constraint.ConstraintLayout>
 </layout>
